get_liver_om_lb_mi_tox_score_list

This function processes liver organ toxicity scores, body weight z-scores, and other related metrics for a set of studies or XPT files. It can output individual scores, z-scores by USUBJID, or averaged scores for multiple studies, and handles errors during the processing steps.

Usage

get_indiv_score_om_lb_mi_domain_df(
  studyid_or_studyids = FALSE,
  path_db,
  fake_study = FALSE,
  use_xpt_file = FALSE,
  all_lb_TESTCD_score = FALSE,
  output_individual_scores = FALSE,
  output_zscore_by_USUBJID = FALSE
)

Arguments

studyid_or_studyids

A character vector or a single study ID to process. If multiple studies are provided, the function processes each study sequentially. (Mandatory)

path_db

A character string specifying the path to the database or directory containing the data files. (Mandatory)

fake_study

A boolean flag indicating if the study data is simulated (TRUE) or real (FALSE). Default is FALSE. (Optional)

use_xpt_file

A boolean flag indicating whether to use an XPT file for the study data. Default is FALSE. (Mandatory)

output_individual_scores

A boolean flag indicating whether individual scores should be returned (TRUE) or averaged scores (FALSE). Default is FALSE. (Optional)

output_zscore_by_USUBJID

A boolean flag indicating whether to output z-scores by USUBJID (TRUE) or averaged scores (FALSE). Default is FALSE. (Optional)

multiple_xpt_folder

A character string specifying the path to the folder containing multiple XPT files. (Optional)

Value

A data frame containing the calculated scores for each study. The type of result depends on the flags passed:

If output_individual_scores is TRUE, a data frame with individual scores for each study is returned.

If output_zscore_by_USUBJID is TRUE, a data frame with z-scores by USUBJID for each study is returned.

If neither flag is set, the function returns a data frame with averaged scores for each study.

Examples

if (FALSE) { # \dontrun{
# Get averaged scores for a single study
result <- get_liver_om_lb_mi_tox_score_list(
  studyid_or_studyids = "Study_001",
  path_db = "path/to/database"
)

# Get individual scores for multiple studies
result_individual_scores <- get_liver_om_lb_mi_tox_score_list(
  studyid_or_studyids = c("Study_001", "Study_002"),
  path_db = "path/to/database",
  output_individual_scores = TRUE
)
} # }